requirements.txt
# Core ML and Computer Vision
opencv-python==4.8.1.78
face-recognition==1.3.0
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.1

# Deep Learning (CPU Only)
torch==2.1.2+cpu
torchvision==0.16.2+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

# Web Framework (for Flask app)
Flask==2.3.3
Werkzeug==2.3.7

# Utilities
psutil==5.9.0
tqdm==4.66.1
h5py==3.10.0
dlib==19.24.2
